Controller,router,middlewares folder payment move to server. folder payment deleted
