--- payment/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "payment",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^1.3.4",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "xendit-node": "^1.21.10"
-  }
-}
