--- payment/routes/index.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const app = express();
-let payment = require("./payment");
-
-app.use("/payments", payment);
-
-module.exports = app;
