+++ server/middlewares/authenticationPayment.js
