+++ server/middlewares/errorHandlerPayment.js
@@ -1,4 +1,4 @@
-let errorHandler = (err, req, res, next) => {
+let errorHandlerPayment = (err, req, res, next) => {
   console.log(err);
   if (err.name === "Bad Request") {
     res.status(400).json({ message: "Amount or Email Customer is Empty" });
@@ -6,4 +6,4 @@ let errorHandler = (err, req, res, next) => {
     res.status(500).json({ message: "Internal Server Error" });
   }
 };
-module.exports = errorHandler;
+module.exports = errorHandlerPayment;
