+++ server/package.json
@@ -22,6 +22,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "mongodb": "5.1",
-    "sequelize": "^6.30.0"
+    "sequelize": "^6.30.0",
+    "xendit-node": "^1.21.10"
   }
 }
